+++ src/arangojs.ts
@@ -1,13 +1,14 @@
+import { aql } from "./aql-query";
+import { CollectionType } from "./collection";
 import { Config } from "./connection";
 import { Database } from "./database";
-import { aql } from "./aql-query";
+import { ArangoError } from "./error";
 
 export default function arangojs(config: Config) {
   return new Database(config);
 }
 
-Object.assign(arangojs, { Database, aql });
+Object.assign(arangojs, { CollectionType, ArangoError, Database, aql });
+export { DocumentCollection, EdgeCollection } from "./collection";
+export { Graph } from "./graph";
 export { Database, aql };
-
-export { DocumentCollection, EdgeCollection } from './collection';
-export { Graph } from './graph';